scripts/docs/requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.3.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-frontmatter>=1.0.0
tqdm>=4.66.0
